fix breack-word setting.
